GCW:1941 Cancel order button changed to link and also added to confirm order page
