+++ app/styles/_confirm.scss
@@ -4,3 +4,13 @@
     font-size: 0.9rem;
   }
 }
+
+.cancel_order {
+  margin-top: -1rem;
+  @media #{$small-only} {
+    margin-right: -4.5rem !important;
+  }
+  @media #{$large-only} {
+    margin-right: -6rem !important;
+  }
+}
