+++ app/styles/_transport_details.scss
@@ -14,6 +14,12 @@
 
   .cancel_order {
     margin-top: -1rem;
+    @media #{$small-only} {
+      margin-right: -4rem !important;
+    }
+    @media #{$large-only} {
+      margin-right: -6rem !important;
+    }
   }
 
   .row {
