+++ app/templates/order/_cancel_order.hbs
@@ -2,8 +2,6 @@
   <div class="small-4 columns">
   </div>
   <div class="small-8 columns cancel_order">
-    <button id="cancel_order" class="button expand drop_off" {{action 'cancelOrderPopUp' order.id}}>
-      {{t "order.cancel_order"}}
-    </button>
+    <a href="#" id="cancel_order" class="drop_off" {{action 'cancelOrderPopUp' order.id}}><u>{{t "order.cancel_order"}}</u></a>
   </div>
 </div>
