+++ app/templates/order/confirm.hbs
@@ -18,6 +18,9 @@
         </button>
       </div>
     </div>
+    {{#if order.isDraft}}
+      {{partial "order/cancel_order"}}
+    {{/if}}
   </div>
   {{partial "cart_content/checkout_full_view"}}
 </div>
